Add option to remove global sidekiq_stats
Sidekiq needs to go through all the queues to get some of the stats being reported
on the sidekiq stats (like the `enqueued`), when you have *a lot* of jobs per second
and a lot of queues this adds way too much load on Redis and kills it.

This adds an option to prevent these global metrics from being reported.